+++ src/test/java/tests/BaseTest.java
@@ -7,7 +7,7 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Listeners;
 import pages.CalendarPage;
-import pages.createWorkout.*;
+import pages.WorkoutPage;
 import steps.LoginSteps;
 import steps.RegistrationSteps;
 import steps.WorkoutSteps;
@@ -23,17 +23,7 @@ public class BaseTest {
     WorkoutSteps workoutSteps;
     RegistrationSteps registrationSteps;
     CalendarPage calendarPage;
-    CreateBikeWorkout createBikeWorkout;
-    CreateCrossTraining createCrossTraining;
-    CreateOtherWorkout createOtherWorkout;
-    CreateRecoveryRehab createRecoveryRehab;
-    CreateRestDay createRestDay;
-    CreateRunWorkout createRunWorkout;
-    CreateStrengthTraining createStrengthTraining;
-    CreateSwimWorkout createSwimWorkout;
-    CreateTransitionWorkout createTransitionWorkout;
-    CreateWalkWorkout createWalkWorkout;
-
+    WorkoutPage workoutPage;
 
     @BeforeMethod
     public void setUp(ITestContext context) {
@@ -48,17 +38,7 @@ public class BaseTest {
         context.setAttribute("driver", driver);
         calendarPage = new CalendarPage(driver);
         loginSteps.openPage();
-
-        createBikeWorkout = new CreateBikeWorkout(driver);
-        createCrossTraining = new CreateCrossTraining(driver);
-        createOtherWorkout = new CreateOtherWorkout(driver);
-        createRecoveryRehab = new CreateRecoveryRehab(driver);
-        createRestDay = new CreateRestDay(driver);
-        createRunWorkout = new CreateRunWorkout(driver);
-        createStrengthTraining = new CreateStrengthTraining(driver);
-        createSwimWorkout = new CreateSwimWorkout(driver);
-        createTransitionWorkout = new CreateTransitionWorkout(driver);
-        createWalkWorkout = new CreateWalkWorkout(driver);
+        workoutPage = new WorkoutPage(driver);
 
     }
